feat: add token common utility
