+++ app/application/usecase/account/login_user.go
@@ -2,19 +2,10 @@ package account
 
 import (
 	"context"
-	"os"
 	"stoneBanking/app/common/utils"
 	"stoneBanking/app/domain/entities/account"
 	customError "stoneBanking/app/domain/errors"
-	"time"
-
-	"github.com/golang-jwt/jwt"
 )
-
-type ClaimStruct struct {
-	jwt.StandardClaims
-	User_id int
-}
 
 func (usecase *usecase) LoginUser(ctx context.Context, loginInput account.Account) (string, error) {
 
@@ -27,29 +18,10 @@ func (usecase *usecase) LoginUser(ctx context.Context, loginInput account.Accoun
 		return "", customError.ErrorAccountLogin
 	}
 
-	token, err := generetateToken(tempAccount)
+	token, err := utils.GenerateToken(string(tempAccount.ExternalID))
 	if err != nil {
 		return "", customError.ErrorAccountTokenGeneration
 	}
 
 	return token, nil
-}
-
-func generetateToken(accountData account.Account) (signedToken string, err error) {
-	mySigningKey := []byte(os.Getenv("SIGN_KEY"))
-	claims := &ClaimStruct{
-		User_id: accountData.ID,
-		StandardClaims: jwt.StandardClaims{
-			ExpiresAt: time.Now().Local().Add(2 * time.Hour).Unix(),
-			Issuer:    "vitorino",
-		},
-	}
-	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
-
-	signedToken, err = token.SignedString(mySigningKey)
-
-	if err != nil {
-		println("Erro na geração do token")
-	}
-	return signedToken, nil
 }
