--- app/common/utils/cpf.go
@@ -1,9 +1,0 @@
-package utils
-
-import "regexp"
-
-func TrimCPF(cpf string) (result string) {
-	regex := regexp.MustCompile("[^0-9]+")
-	result = regex.ReplaceAllString(cpf, "")
-	return result
-}
